pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster-dbt>=0.28.1",
    "dagster-dlt>=0.28.1",
    "dagster-duckdb>=0.28.1",
    "dagster-webserver>=1.12.1",
    "dbt-core>=1.10.13",
    "dbt-duckdb>=1.10.0",
    "dbt-postgres>=1.9.1",
    "dlt[duckdb,postgres]>=1.18.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
